# openapi.yaml
openapi: 3.1.0
info:
  title: RGB Lightning Node
  description: |-
    This is the OpenAPI specification for the
    [RGB Lightning Node](https://github.com/RGB-Tools/rgb-lightning-node) APIs.
  license:
    name: MIT
    url: https://mit-license.org/
  version: 0.1.0
servers:
  - url: http://localhost:3001
  - url: http://localhost:3002
  - url: http://localhost:3003
...
paths:
...
  /apay/new:
    post:
      tags:
        - Async Payments
      summary: Send a new order with async-payment hashes
      description: Derive deterministic payment hashes for async payments from the unlocked wallet. Send an async_order.new JSON-RPC request to the connected invoice-host peer so it can track those hashes and process asynchronous payments. A peer connection alone is not sufficient; the recipient-client must already have a live channel with the invoice-host peer. 
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsyncOrderNewRequest'
      responses:
        '200':
          description: Host accepted the async order request and returned order state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOrderNewResponse'
  /apay/outboundinvoice:
    post:
      tags:
        - Async Payments
      summary: Request an outbound LN invoice
      description: Request an outbound invoice from the connected recipient-client peer for an async payment claim session. A peer connection alone is not enough; the invoice-host must already have a live channel with the recipient-client peer.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsyncOrderOutboundInvoiceRequest'
      responses:
        '200':
          description: Recipient peer returned an outbound invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOrderOutboundInvoiceResponse'
  /assetbalance:
    post:
      tags:
        - RGB
      summary: Get the balance of an asset
      description: Get the balance for the provided RGB asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetBalanceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetBalanceResponse'
  /assetlink:
    post:
      tags:
        - RGB
      summary: Create a contract link
      description: Link a parent IFA contract to a child IFA contract using a link-right UTXO
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetLinkRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetLinkResponse'
  /assetmetadata:
    post:
      tags:
        - RGB
      summary: Get the metadata of an asset
      description: Get the metadata for the provided RGB asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetMetadataRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetMetadataResponse'
  /backup:
    post:
      tags:
        - Other
      summary: Backup the node
      description: Create a backup of the node's data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /btcbalance:
    post:
      tags:
        - On-chain
      summary: Get the BTC balance
      description: Get the node's bitcoin balance for the vanilla and colored wallets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BtcBalanceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BtcBalanceResponse'
  /cancelhodlinvoice:
    post:
      tags:
        - Invoices
      summary: Cancel a HODL invoice
      description: Cancel a held HTLC for a HODL invoice. Rejects cancellation if a settlement is already in progress.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCancelRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /changepassword:
    post:
      tags:
        - Other
      summary: Change the password
      description: Change the node's password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /checkindexerurl:
    post:
      tags:
        - Other
      summary: Check an indexer URL
      description: Check the given indexer URL is valid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckIndexerUrlRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIndexerUrlResponse'
  /checkproxyendpoint:
    post:
      tags:
        - Other
      summary: Check a proxy endpoint
      description: Check the given proxy endpoint is valid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckProxyEndpointRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /claimhodlinvoice:
    post:
      tags:
        - Invoices
      summary: Claim a HODL invoice
      description: Claim a held HTLC for a HODL invoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimHodlInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimHodlInvoiceResponse'
  /closechannel:
    post:
      tags:
        - Channels
      summary: Close a channel
      description: Close a LN channel cooperatively or forcibly
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseChannelRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /connectpeer:
    post:
      tags:
        - Peers
      summary: Connect to a peer
      description: Connect to the provided LN peer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectPeerRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /createutxos:
    post:
      tags:
        - RGB
      summary: Create UTXOs
      description: Create UTXOs to be used for RGB operations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUtxosRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /decodelninvoice:
    post:
      tags:
        - Invoices
      summary: Decode a LN invoice
      description: Decode the provided LN invoice string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeLNInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeLNInvoiceResponse'
  /decodergbinvoice:
    post:
      tags:
        - RGB
      summary: Decode an RGB invoice
      description: Decode the provided RGB invoice string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeRGBInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeRGBInvoiceResponse'
  /decodeswapstring:
    post:
      tags:
        - Swaps
      summary: Decode a swapstring
      description: Decode the provided swapstring
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecodeSwapstringRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeSwapstringResponse'
  /disconnectpeer:
    post:
      tags:
        - Peers
      summary: Disconnect from a peer
      description: Disconnect from the provided LN peer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisconnectPeerRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /estimatefee:
    post:
      tags:
        - On-chain
      summary: Get fee estimation
      description: Get on-chain fee estimation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateFeeRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateFeeResponse'
  /failtransfers:
    post:
      tags:
        - RGB
      summary: Fail RGB transfers
      description: Set the status for eligible RGB transfers to
        `TransferStatus::Failed`.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailTransfersRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailTransfersResponse'
  /getassetmedia:
    post:
      tags:
        - RGB
      summary: Get an asset media
      description: Get the hex string of the media bytes of the provided media digest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAssetMediaRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAssetMediaResponse'
  /getchannelid:
    post:
      tags:
        - Channels
      summary: Get a channel's ID
      description: Get a channel's ID from its former temporary channel ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetChannelIdRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetChannelIdResponse'
  /getpayment:
    post:
      tags:
        - Payments
      summary: Get a payment
      description: Get a payment by its payment hash
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPaymentRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPaymentResponse'
  /getswap:
    post:
      tags:
        - Swaps
      summary: Get a swap
      description: Get a swap by its payment hash
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetSwapRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSwapResponse'
  /inflate:
    post:
      tags:
        - RGB
      summary: Inflate RGB assets
      description: Inflate RGB assets on-chain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InflateRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InflateResponse'
  /init:
    post:
      tags:
        - Other
      summary: Init the node
      description: Initialize a new node, optionally providing a mnemonic.
        If no mnemonic is provided a new one is generated.
        Note that a mnemonic alone is not sufficient to recover RGB state, which also requires client-side data.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitResponse'
  /invoicestatus:
    post:
      tags:
        - Invoices
      summary: Get an invoice status
      description: Get the status of the provided LN invoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceStatusRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceStatusResponse'
  /issueassetcfa:
    post:
      tags:
        - RGB
      summary: Issue an RGB CFA asset
      description: Issue an RGB CFA asset. To provide a media first call the /postassetmedia API.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetCFARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetCFAResponse'
  /issueassetifa:
    post:
      tags:
        - RGB
      summary: Issue an RGB IFA asset
      description: Issue an RGB IFA asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetIFARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetIFAResponse'
  /issueassetnia:
    post:
      tags:
        - RGB
      summary: Issue an RGB NIA asset
      description: Issue an RGB NIA asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetNIARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetNIAResponse'
  /issueassetuda:
    post:
      tags:
        - RGB
      summary: Issue an RGB UDA asset
      description: Issue an RGB UDA asset. To provide a media first call the /postassetmedia API.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAssetUDARequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAssetUDAResponse'
  /keysend:
    post:
      tags:
        - Payments
      summary: Send to a peer spontaneously
      description: Send bitcoins and RGB assets to a LN peer spontaneously (without a LN invoice)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeysendRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeysendResponse'
  /listassets:
    post:
      tags:
        - RGB
      summary: List assets
      description: List the node's RGB assets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListAssetsRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAssetsResponse'
...
  /listpayments:
    get:
      tags:
        - Payments
      summary: List payments
      description: >-
        List the node's LN payments (inbound and outbound) ordered most-recent
        first. Supports index-based cursor pagination: pass the
        last_index_offset of a page as the next index_offset to fetch the next,
        older page. Optionally filter by status, direction (inbound/outbound)
        and creation time interval.
      parameters:
        - in: query
          name: index_offset
          required: false
          schema:
            type: integer
            format: uint64
          description: >-
            Exclusive upper-bound cursor: only payments with a lower index are
            returned. 0 or absent means start from the most recent payment.
        - in: query
          name: max_payments
          required: false
          schema:
            type: integer
            format: uint64
          description: >-
            Maximum number of payments to return. Defaults to 100 when absent or
            zero.
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum:
              - Pending
              - Succeeded
              - Failed
              - Claimable
              - Claiming
              - Cancelled
          description: Return only payments with this HTLC status.
        - in: query
          name: direction
          required: false
          schema:
            type: string
            enum:
              - Inbound
              - Outbound
          description: Return only inbound or only outbound payments.
        - in: query
          name: created_after
          required: false
          schema:
            type: integer
            format: uint64
          description: >-
            Return only payments created at or after this Unix timestamp
            (seconds).
        - in: query
          name: created_before
          required: false
          schema:
            type: integer
            format: uint64
          description: >-
            Return only payments created at or before this Unix timestamp
            (seconds).
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPaymentsResponse'
...
  /listtransactions:
    post:
      tags:
        - On-chain
      summary: List transactions
      description: >-
        List the node's on-chain transactions, ordered most-recent first.
        Supports index-based cursor pagination: pass the last_index_offset of a
        page as the next index_offset to fetch the next, older page.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListTransactionsRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTransactionsResponse'
  /listtransfers:
    post:
      tags:
        - RGB
      summary: List transfers
      description: >-
        List the node's RGB transfers for an asset, ordered most-recent first.
        Supports index-based cursor pagination (idx) and optional filtering by
        status and creation time interval.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListTransfersRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTransfersResponse'
  /listunspents:
    post:
      tags:
        - On-chain
      summary: List unspents
      description: >-
        List the unspent outputs of the internal BDK wallet. Supports
        index-based cursor pagination: pass the last_index_offset of a page as
        the next index_offset to fetch the next page.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListUnspentsRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUnspentsResponse'
...
  /lninvoice:
    post:
      tags:
        - Invoices
      summary: Get a LN invoice
      description: >-
        Get a LN invoice to receive a payment. Provide `payment_hash` to create a HODL invoice.
        Provide `description_hash` to include an out-of-band BOLT11 description hash.
        Provide `min_final_cltv_expiry_delta` to request an explicit inbound final CLTV policy.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LNInvoiceRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LNInvoiceResponse'
...
components:
  schemas:
...
    AsyncOrderNewRequest:
      type: object
      required:
        - host_node_id
      dependentRequired:
        username:
          - domain
        domain:
          - username
      properties:
        host_node_id:
          type: string
          example: 02123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0
        username:
          type: string
          description: >-
            Optional. LSP-assigned Lightning Address localpart. When supplied together with `domain`,
            the recipient signs a `recipient_pubkey <-> username@domain` address attestation and includes it with the order.
            Source it from the LSP's `GET /lightning_address/by_pubkey/{pubkey}` so it matches what the LSP serves.
          example: xalkan
        domain:
          type: string
          description: >-
            Optional. Lightning Address domain; required together with `username` to produce the address attestation.
          example: utexo.com
...
    AsyncOrderOutboundInvoiceRequest:
      type: object
      required:
        - client_node_id
        - params
      properties:
        client_node_id:
          type: string
          example: 02270dadcd6e7ba0ef707dac72acccae1a3607453a8dd2aef36ff3be4e0d31f043
        params:
          $ref: '#/components/schemas/AsyncOrderRequestInvoiceParams'
...
    EmptyResponse:
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Biscuit
security:
  - bearerAuth: []